refactor: transforms hydrate main menu into an use case
